+++ CHANGES.md
@@ -1,9 +1,12 @@
-# Version 0.2.0
+# Version 0.2.0, 2020-05-11
 
 - Compatibility with OCaml 4.11.0
 
 - More utility functions in Metapp module to construct and destruct OCaml's
   parsetree in a version-independent way
+
+- Functions for invoking OCaml compiler and loading modules dynamically are
+  exported in the package `metapp.dyncompile`.
 
 # Version 0.1.0, 2020-02-27